requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
loguru
twilio
livekit-api
deepgram-sdk
pipecat
python-multipart
